Seperate CWL generation logic from application parsing logic
Cloases #15
