+++ app_pack_generator/__init__.py
@@ -1,5 +1,6 @@
-from .application import ApplicationNotebook
 from .git import GitManager, GitRepoError
 from .docker import DockerUtil
+from .application import ApplicationNotebook
+from .process import CWL, Descriptor
 
 from .version import __version__
