+++ app_pack_generator/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.2.1"
+__version__ = "0.2.2"
